requirements.txt
Requests==2.30.0
streamlit==1.22.0
Streamlit_Fullcalendar_Component==0.0.1
timeblok_py==0.4.5
openai==0.27.4